test: moving proxy proto
