+++ source/extensions/filters/listener/proxy_protocol/BUILD
@@ -39,10 +39,6 @@ envoy_cc_library(
 envoy_cc_extension(
     name = "config",
     srcs = ["config.cc"],
-    # TODO(#9953) clean up.
-    extra_visibility = [
-        "//test/integration:__subpackages__",
-    ],
     deps = [
         "//envoy/registry",
         "//envoy/server:filter_config_interface",
