+++ test/extensions/filters/listener/proxy_protocol/BUILD
@@ -59,3 +59,25 @@ envoy_cc_fuzz_test(
         "//test/extensions/filters/listener/common/fuzz:listener_filter_fuzzer_lib",
     ],
 )
+
+envoy_extension_cc_test(
+    name = "proxy_proto_integration_test",
+    srcs = [
+        "proxy_proto_integration_test.cc",
+        "proxy_proto_integration_test.h",
+    ],
+    extension_names = ["envoy.filters.listener.proxy_protocol"],
+    deps = [
+        "//source/common/buffer:buffer_lib",
+        "//source/common/http:codec_client_lib",
+        "//source/extensions/access_loggers/file:config",
+        "//source/extensions/filters/listener/proxy_protocol:config",
+        "//source/extensions/filters/network/tcp_proxy:config",
+        "//test/integration:http_integration_lib",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/config/bootstrap/v3:pkg_cc_proto",
+        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/access_loggers/file/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/filters/network/tcp_proxy/v3:pkg_cc_proto",
+    ],
+)
